Implement update quantity
Currently shopping cart updates with the current quantity as the default in the
dropdown selector. Updating changes the quantity to the value, not increments it,
so currently a max of 15 is the max, but can be easily changed by just changing
one piece of code.

The params are coming in a little funny. See the patch method in the cart controller
if you want to see how it works.