Change order status to enums model, implement predicate methods
We wanted to change our order status attribute to have a default behavior, and to
use enums for consistency and ease.Was a seemingly simple refactor, but involved
a migration to create the default behavior. Now all 4 order statuses are defined
in the enum model, with "ordered" mapping to the default value.

Tests were refactored to remove creation with strings
and you will want to note that in creating any future orders you need to specify
the number corresponding to the status you would like. However, this does let
us simply ask if something is .ordered? or cancelled? Hence, I also had to
refactor a couple of methods and views.

Additionally, the other issues in this waffle card were changed, including
switching the order of has_ and has_through attributes on classes and adding
better Capybara CSS matchers on a couple of tests.