Add charity functionality and bugfix tag routes
This is a big commit for charity functionality, though is not currently complete,
either in styling or in passing tests.

We will need to address an inconsistency in how gif creation is handled at the
controller/router level, where some paths are handed off to admin, and others are not.
This seemed to be working but became confusing when I tried to add the addition
of selecting a charity and was not hitting the controller I thought I should be.

Pieces finished:
* Charity creation, at the router/controller/view level
* Seeding our selected charities into dev
* Styling of a charity index that shows all 3 charities with links to view the
gifs associated with them
* Quick setup of a charity show view that lists all gifs. This is linked from the
gif show page, and seems like it should have all the charity info or else it's a
little misleading, but will take some more styling to incorporate that

This commit also has a bugfix for the tag bug we had experienced, stemming from
incorrect routing of the tag vanity urls to `/tagname` rather than `/tag/tagname`.